Change long to int everywhere

+++ src/gameObject/Animation.cpp
@@ -17,13 +17,13 @@ sf::IntRect Animation::getNextFrame() {
     return this->animationFrames.at(currentFrame);
 }
 
-void Animation::addAnimationFrame(long x, long y, long width, long height) {
+void Animation::addAnimationFrame(int x, int y, int width, int height) {
     this->animationFrames.push_back(sf::IntRect(x, y, width, height));
 }
 
-void Animation::addAnimationFrames(long fromX, long fromY, long toX, long toY, long width, long height) {
-    for (long y = fromY; y <= toY; y += height) {
-        for (long x = fromX; x <= textureWidth && !(x > toX && y >= toY); x+= width) {
+void Animation::addAnimationFrames(int fromX, int fromY, int toX, int toY, int width, int height) {
+    for (int y = fromY; y <= toY; y += height) {
+        for (int x = fromX; x <= textureWidth && !(x > toX && y >= toY); x+= width) {
             this->animationFrames.push_back(sf::IntRect(x, y, width, height));
         }
     }
